Renamed data layer model classes
