+++ src/main/java/uk/gov/hmcts/reform/bulkscanprocessor/entity/reports/ZipFileSummary.java
@@ -2,7 +2,7 @@ package uk.gov.hmcts.reform.bulkscanprocessor.entity.reports;
 
 import java.time.Instant;
 
-public interface ZipFileSummaryItem {
+public interface ZipFileSummary {
 
     String getZipFileName();
